refactor(training): api.py → pacote api/ com 12 sub-routers (Fases 1.3–1.8)
- Fase 1.3: mappers.py — extrai mapeadores de schema (282 linhas)
- Fase 1.4: sessions.py (11 handlers), blocks.py (6), attendance.py (2),
  wellness.py (6), planning.py (8), execution.py (5), feedback.py (3),
  attention.py (4), recommendations.py (3), eligibility.py (2),
  analytics.py (1), chat.py (2) — 53 handlers migrados
- Fase 1.5: _shared.py removido; deps.py e errors.py canônicos
- Fase 1.6: __init__.py vira thin aggregator (add_router x12)
- Fase 1.7: conftest.py + test_route_inventory.py atualizados
- Fase 1.8: gate fixes — endpoints.yaml (53 runtime_handler_ref),
  module_manifest.yaml, source graph + context bundle regenerados,
  check_architecture_docs.py aceita api/, codegen parity agrega sub-arquivos

Gates: 287 passed, 0 failed (training + pipeline_gates + parity)
