+++ _reports/READINESS_DASHBOARD.md
@@ -1,5 +1,5 @@
 # Dashboard de Readiness - HB Track
-> Gerado em 2026-04-21T07:44:00Z | run_id: `20260421T074400_ac4a4a` | health: **100/100** | overall: **PASS**
+> Gerado em 2026-04-21T21:26:38Z | run_id: `20260421T212638_31d792` | health: **97/100** | overall: **FAIL**
 
 ## Modulos
 
@@ -70,7 +70,7 @@
 | DEPLOY_READINESS_GATE | PASS | nao |
 | DATA_MIGRATION_GATE | PASS | nao |
 | MONITORING_POLICY_GATE | PASS | nao |
-| HANDOFF_COHERENCE_GATE | PASS | sim |
+| HANDOFF_COHERENCE_GATE | FAIL | sim |
 | MODULE_STATUS_COHERENCE_GATE | PASS | sim |
 | SURFACE_PROMOTION_COHERENCE_GATE | PASS | sim |
 | CROSS_MODULE_BOUNDARY_GATE | PASS | nao |
@@ -88,4 +88,4 @@
 | DOC_USAGE_GATE | PASS | sim |
 | CANON_CONTRACT_DRIVEN_PARITY_GATE | PASS | sim |
 | HBTRACK_CANON_PARITY_GATE | PASS | sim |
-| READINESS_SUMMARY_GATE | PASS | nao |
+| READINESS_SUMMARY_GATE | FAIL | nao |
