+++ _reports/contract_gates/precommit.latest.json
@@ -1,6 +1,6 @@
 {
   "pipeline_id": "HB_TRACK_CONTRACT_GATES",
-  "timestamp_utc": "2026-04-21T10:07:57Z",
+  "timestamp_utc": "2026-04-21T19:08:12Z",
   "target": {
     "scope": "system",
     "module": null,
@@ -9,7 +9,7 @@
     "workflow_scope": "/home/davis/HB-TRACK/contracts/workflows"
   },
   "environment": {
-    "git_commit": "4c82d911",
+    "git_commit": "b5b38a74",
     "python_version": "3.12.3",
     "tool_versions": {
       "redocly": "1.34.10",
@@ -31,7 +31,7 @@
   "report_artifacts": {
     "scoped_report_path": "/home/davis/HB-TRACK/_reports/contract_gates/precommit.latest.json",
     "canonical_report_path": "/home/davis/HB-TRACK/_reports/contract_gates/latest.json",
-    "run_dir": "/home/davis/HB-TRACK/_reports/runs/20260421T100757_dce5d1"
+    "run_dir": "/home/davis/HB-TRACK/_reports/runs/20260421T190812_2e7124"
   },
   "status_detail": {
     "active_gates_passed": 24,
@@ -261,7 +261,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 43
+        "duration_ms": 60
       }
     },
     {
@@ -286,7 +286,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 6063
+        "duration_ms": 1320
       }
     },
     {
@@ -397,7 +397,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 17
+        "duration_ms": 16
       }
     },
     {
@@ -1018,7 +1018,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 81
+        "duration_ms": 48
       }
     },
     {
@@ -1104,7 +1104,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 734
+        "duration_ms": 735
       }
     },
     {
@@ -1127,7 +1127,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 2227
+        "duration_ms": 1329
       }
     },
     {
@@ -1149,7 +1149,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 3167
+        "duration_ms": 1230
       }
     },
     {
@@ -1188,7 +1188,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 580
+        "duration_ms": 566
       }
     },
     {
@@ -4140,7 +4140,7 @@
         "errors": 0,
         "warnings": 144,
         "violations": 144,
-        "duration_ms": 2198
+        "duration_ms": 1759
       }
     },
     {
@@ -4208,7 +4208,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 13
+        "duration_ms": 2
       }
     },
     {
@@ -4823,7 +4823,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 1760
+        "duration_ms": 1682
       }
     },
     {
@@ -4899,7 +4899,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 1802
+        "duration_ms": 1475
       }
     },
     {
@@ -4942,7 +4942,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 58
+        "duration_ms": 60
       }
     },
     {
@@ -4964,7 +4964,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 1351
+        "duration_ms": 1206
       }
     },
     {
@@ -5020,7 +5020,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 64
+        "duration_ms": 59
       }
     },
     {
@@ -5040,7 +5040,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 6
+        "duration_ms": 3
       }
     },
     {
@@ -5121,7 +5121,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 638
+        "duration_ms": 233
       }
     },
     {
@@ -5319,7 +5319,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 30
+        "duration_ms": 8
       }
     },
     {
@@ -5341,7 +5341,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 188
+        "duration_ms": 161
       }
     },
     {
@@ -5423,7 +5423,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 10
+        "duration_ms": 6
       }
     },
     {
@@ -5445,7 +5445,7 @@
         "errors": 0,
         "warnings": 0,
         "violations": 0,
-        "duration_ms": 21
+        "duration_ms": 17
       }
     },
     {
