+++ _reports/evidence/module_readiness_scorecard.json
@@ -1,7 +1,7 @@
 {
   "artifact_id": "HBTRACK_MODULE_READINESS_SCORECARD",
-  "generated_at_utc": "2026-04-21T07:44:00Z",
-  "overall_status": "PASS",
+  "generated_at_utc": "2026-04-21T21:26:38Z",
+  "overall_status": "FAIL",
   "source_report": "_reports/contract_gates/latest.json",
   "modules": [
     {
