+++ _reports/pipeline_health.json
@@ -1,11 +1,11 @@
 {
-  "run_id": "20260421T074400_ac4a4a",
-  "timestamp_utc": "2026-04-21T07:44:00Z",
-  "health_score": 100,
+  "run_id": "20260421T212638_31d792",
+  "timestamp_utc": "2026-04-21T21:26:38Z",
+  "health_score": 97,
   "gates_total": 62,
-  "gates_passed": 62,
-  "gates_failed": 0,
-  "blocking_fails": 0,
-  "overall_status": "PASS",
-  "exit_code": 0
+  "gates_passed": 60,
+  "gates_failed": 2,
+  "blocking_fails": 1,
+  "overall_status": "FAIL",
+  "exit_code": 2
 }
