+++ _reports/pipeline_history.jsonl
@@ -1620,3 +1620,5 @@
 {"run_id": "20260421T074054_81351e", "timestamp_utc": "2026-04-21T07:40:54Z", "overall_status": "FAIL", "exit_code": 2, "health_score": 97, "git_commit": "3223548d"}
 {"run_id": "20260421T074255_036f1f", "timestamp_utc": "2026-04-21T07:42:55Z", "overall_status": "PASS", "exit_code": 0, "health_score": 100, "git_commit": "3223548d"}
 {"run_id": "20260421T074400_ac4a4a", "timestamp_utc": "2026-04-21T07:44:00Z", "overall_status": "PASS", "exit_code": 0, "health_score": 100, "git_commit": "3223548d"}
+{"run_id": "20260421T212522_47fe70", "timestamp_utc": "2026-04-21T21:25:22Z", "overall_status": "FAIL", "exit_code": 2, "health_score": 95, "git_commit": "29132414"}
+{"run_id": "20260421T212638_31d792", "timestamp_utc": "2026-04-21T21:26:38Z", "overall_status": "FAIL", "exit_code": 2, "health_score": 97, "git_commit": "29132414"}
