+++ src/training/api/analytics.py
@@ -1,0 +1,61 @@
+"""analytics sub-router — training.api.
+
+Endpoints:
+  GET  /training-sessions/{id}/load-chart
+"""
+
+import uuid
+
+from ninja import Router
+from ninja.errors import HttpError
+
+from ..application.use_cases import (
+    GetLoadChartInput,
+    GetLoadChartUseCase,
+)
+from ..infrastructure.repository import (
+    ExecutionRecordRepository,
+    TrainingSessionRepository,
+)
+from ..schemas import ErrorOut, LoadChartEntryOut, LoadChartOut
+from .deps import _get_actor_role
+from .errors import map_exceptions
+
+router = Router()
+
+
+@router.get(
+    "/training-sessions/{id}/load-chart",
+    response={200: LoadChartOut, 401: ErrorOut, 403: ErrorOut, 404: ErrorOut},
+)
+@map_exceptions
+def get_load_chart(request, id: uuid.UUID):
+    session_repo = TrainingSessionRepository()
+    execution_record_repo = ExecutionRecordRepository()
+    result = GetLoadChartUseCase(session_repo, execution_record_repo).execute(
+        GetLoadChartInput(
+            session_id=id,
+            actor_role=_get_actor_role(request),
+        )
+    )
+    return 200, LoadChartOut(
+        session_id=result.session.id,
+        planned_load=result.session.planned_load,
+        actual_load_recorded=result.session.actual_load_recorded,
+        entries=[
+            LoadChartEntryOut(
+                id=e.id,
+                recorded_at=e.recorded_at,
+                planned_value=e.planned_value,
+                actual_value=e.actual_value,
+                planned_unit=e.planned_unit,
+                actual_unit=e.actual_unit,
+                notes=e.notes,
+            )
+            for e in result.load_entries
+        ],
+    )
+
+
+def register(parent: Router) -> None:
+    parent.add_router("", router)
